Fix #472
