+++ Environment_oM/Elements/Opening.cs
@@ -26,7 +26,7 @@ using BH.oM.Geometry;
 using BH.oM.Base;
 using BH.oM.Geometry;
 using BH.oM.Analytical.Elements;
-using BH.oM.Environment.Properties;
+using BH.oM.Environment.Fragments;
 using BH.oM.Physical.Constructions;
 
 namespace BH.oM.Environment.Elements
