+++ Environment_oM/Environment_oM.csproj
@@ -62,23 +62,23 @@
     <Compile Include="Gains\Gain.cs" />
     <Compile Include="Gains\IGain.cs" />
     <Compile Include="MaterialFragments\Enums\Function.cs" />
-    <Compile Include="Properties\Fragment\BuildingAnalyticalFragment.cs" />
-    <Compile Include="Properties\Fragment\BuildingContextFragment.cs" />
-    <Compile Include="Properties\Fragment\BuildingResultFragment.cs" />
-    <Compile Include="Properties\Fragment\CoefficientFragment.cs" />
-    <Compile Include="Properties\Fragment\Enums\SizingMethod.cs" />
-    <Compile Include="Properties\Fragment\EnvironmentConstructionFragment.cs" />
-    <Compile Include="Properties\Fragment\LoadFragment.cs" />
-    <Compile Include="Properties\Fragment\OriginContextFragment.cs" />
-    <Compile Include="Properties\Fragment\PanelAnalyticalFragment.cs" />
-    <Compile Include="Properties\Fragment\PanelContextFragment.cs" />
-    <Compile Include="Properties\Fragment\RadiationFragment.cs" />
-    <Compile Include="Properties\Fragment\SpaceAnalyticalFragment.cs" />
-    <Compile Include="Properties\Fragment\SpaceContextFragment.cs" />
-    <Compile Include="Properties\IGainProperties.cs" />
+    <Compile Include="Fragments\BuildingAnalyticalFragment.cs" />
+    <Compile Include="Fragments\BuildingContextFragment.cs" />
+    <Compile Include="Fragments\BuildingResultFragment.cs" />
+    <Compile Include="Fragments\CoefficientFragment.cs" />
+    <Compile Include="Fragments\Enums\SizingMethod.cs" />
+    <Compile Include="Fragments\EnvironmentConstructionFragment.cs" />
+    <Compile Include="Fragments\LoadFragment.cs" />
+    <Compile Include="Fragments\OriginContextFragment.cs" />
+    <Compile Include="Fragments\PanelAnalyticalFragment.cs" />
+    <Compile Include="Fragments\PanelContextFragment.cs" />
+    <Compile Include="Fragments\RadiationFragment.cs" />
+    <Compile Include="Fragments\SpaceAnalyticalFragment.cs" />
+    <Compile Include="Fragments\SpaceContextFragment.cs" />
+    <Compile Include="Gains\IGainProperties.cs" />
     <Compile Include="Gains\InternalCondition.cs" />
-    <Compile Include="Properties\InfiltrationGain.cs" />
-    <Compile Include="Properties\LatentEquipmentGain.cs" />
+    <Compile Include="Gains\InfiltrationGain.cs" />
+    <Compile Include="Gains\LatentEquipmentGain.cs" />
     <Compile Include="Gains\Profile.cs" />
     <Compile Include="Gains\Thermostat.cs" />
     <Compile Include="IEnvironmentObject.cs" />
@@ -90,11 +90,11 @@
     <Compile Include="MaterialFragments\GasMaterial.cs" />
     <Compile Include="MaterialFragments\IEnvironmentMaterial.cs" />
     <Compile Include="MaterialFragments\SolidMaterial.cs" />
-    <Compile Include="Properties\Fragment\IBHoMFragment.cs" />
-    <Compile Include="Properties\LightingGain.cs" />
-    <Compile Include="Properties\PeopleGain.cs" />
-    <Compile Include="Properties\PollutantGain.cs" />
-    <Compile Include="Properties\SensibleEquipmentGain.cs" />
+    <Compile Include="Fragments\IBHoMFragment.cs" />
+    <Compile Include="Gains\LightingGain.cs" />
+    <Compile Include="Gains\PeopleGain.cs" />
+    <Compile Include="Gains\PollutantGain.cs" />
+    <Compile Include="Gains\SensibleEquipmentGain.cs" />
     <Compile Include="Results\Enums\ProfileResultType.cs" />
     <Compile Include="Results\Enums\ProfileResultUnit.cs" />
     <Compile Include="Results\Enums\SimulationResultType.cs" />
