+++ Environment_oM/Fragments/PanelAnalyticalFragment.cs
@@ -28,11 +28,34 @@ using System.Threading.Tasks;
 
 using BH.oM.Base;
 
-namespace BH.oM.Environment.Properties
+namespace BH.oM.Environment.Fragments
 {
-    public class BuildingResultFragment : BHoMObject, IBHoMFragment
+    public class PanelAnalyticalFragment : BHoMObject, IBHoMFragment
     {
-        public double PeakCooling { get; set; } = 0;
-        public double PeakHeating { get; set; } = 0;
+        public double Altitude { get; set; } = 0.0;
+        public double AltitudeRange { get; set; } = 0.0;
+        public double Inclination { get; set; } = 0.0;
+        public double InclinationRange { get; set; } = 0.0;
+        public double Orientation { get; set; } = 0.0;
+        public double GValue { get; set; } = 0.0;
+        public double LTValue { get; set; } = 0.0;
+        public double UValue { get; set; } = 0.0;
+
+        public double ApertureFlowIn { get; set; } = 0.0;
+        public double ApertureFlowOut { get; set; } = 0.0;
+        public double ApertureOpening { get; set; } = 0.0;
+        public double ExternalCondensation { get; set; } = 0.0;
+        public double ExternalConduction { get; set; } = 0.0;
+        public double ExternalConvection { get; set; } = 0.0;
+        public double ExternalLongWave { get; set; } = 0.0;
+        public double ExternalSolar { get; set; } = 0.0;
+        public double ExternalTemperature { get; set; } = 0.0;
+        public double InternalCondensation { get; set; } = 0.0;
+        public double InternalConduction { get; set; } = 0.0;
+        public double InternalConvection { get; set; } = 0.0;
+        public double InternalLongWave { get; set; } = 0.0;
+        public double InternalSolar { get; set; } = 0.0;
+        public double InternalTemperature { get; set; } = 0.0;
+        public double InterstitialCondensation { get; set; } = 0.0;
     }
 }
