+++ Environment_oM/Fragments/PanelContextFragment.cs
@@ -14,21 +14,27 @@
  * The BHoM is distributed in the hope that it will be useful,              
  * but WITHOUT ANY WARRANTY; without even the implied warranty of               
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
- * GNU Lesser General Public License for more details.                          
+ * GNU Lesser General Public License for more details.
  *                                                                            
  * You should have received a copy of the GNU Lesser General Public License     
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-namespace BH.oM.Environment.Properties
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using BH.oM.Base;
+
+namespace BH.oM.Environment.Fragments
 {
-    public enum SizingMethod
+    public class PanelContextFragment : BHoMObject, IBHoMFragment
     {
-        Undefined,
-        DesignSizingOnly,
-        LoadPerArea,
-        Mixed,
-        NoSizing,
-        Sizing,
+        public bool IsAir { get; set; } = false;
+        public bool IsGround { get; set; } = false;
+        public string Colour { get; set; } = "";
+        public bool Reversed { get; set; } = false;
     }
 }
