+++ Environment_oM/Gains/IGainProperties.cs
