+++ Environment_oM/Gains/InfiltrationGain.cs
@@ -14,7 +14,7 @@
  * The BHoM is distributed in the hope that it will be useful,              
  * but WITHOUT ANY WARRANTY; without even the implied warranty of               
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
- * GNU Lesser General Public License for more details.
+ * GNU Lesser General Public License for more details.                          
  *                                                                            
  * You should have received a copy of the GNU Lesser General Public License     
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
@@ -28,11 +28,12 @@ using System.Threading.Tasks;
 
 using BH.oM.Base;
 
-namespace BH.oM.Environment.Properties
+namespace BH.oM.Environment.Gains
 {
-    public class LoadFragment : BHoMObject, IBHoMFragment
+    public class InfiltrationGain : BHoMObject, IGainProperties
     {
-        public double HeatingLoad { get; set; } = 0.0;
-        public double CoolingLoad { get; set; } = 0.0;
+        public Profile Profile { get; set; } = new Profile();
+        public GainUnit Unit { get; set; } = GainUnit.Undefined;
+        public double Value { get; set; } = 0.0;
     }
 }
