+++ Environment_oM/Gains/LightingGain.cs
@@ -14,7 +14,7 @@
  * The BHoM is distributed in the hope that it will be useful,              
  * but WITHOUT ANY WARRANTY; without even the implied warranty of               
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
- * GNU Lesser General Public License for more details.
+ * GNU Lesser General Public License for more details.                          
  *                                                                            
  * You should have received a copy of the GNU Lesser General Public License     
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
@@ -28,12 +28,15 @@ using System.Threading.Tasks;
 
 using BH.oM.Base;
 
-namespace BH.oM.Environment.Properties
+namespace BH.oM.Environment.Gains
 {
-    public class CoefficientFragment : BHoMObject, IBHoMFragment
+    public class LightingGain : BHoMObject, IGainProperties
     {
-        public double LightingViewCoefficient { get; set; } = 0.49;
-        public double OccupantViewCoefficient { get; set; } = 0.227;
-        public double EquipmentViewCoefficient { get; set; } = 0.372;
+        public Profile Profile { get; set; } = new Profile();
+        public GainUnit Unit { get; set; } = GainUnit.Undefined;
+        public double RadiantFraction { get; set; } = 0.0;
+        public double ViewCoefficient { get; set; } = 0.0;
+        public double LuminousEfficacy { get; set; } = 0.0;
+        public double Value { get; set; } = 0.0;
     }
 }
