+++ Environment_oM/MaterialFragments/Absorptance.cs
@@ -27,7 +27,7 @@ using System.Text;
 using System.Threading.Tasks;
 
 using BH.oM.Base;
-using BH.oM.Environment.Properties;
+using BH.oM.Environment.Fragments;
 
 namespace BH.oM.Environment.MaterialFragments
 {
